requirements.txt
streamlit
gradio
ultralytics
opencv-python-headless
gdown
Pillow
